Put tests in one file to avoid duplication.
