+++ test/int/graphql.test.ts
@@ -1,10 +1,11 @@
-// graphql.test.ts
+// TransferSummary.test.ts
 import { ObjectId } from 'mongodb';
 import { getMongoClient, Collection, closeMongoClientConnection } from '../../src/lib/mongo';
+import TransferSummary from '../../src/schema/TransferSummary';
 import { objectType, extendType, queryField } from 'nexus';
 const testTransfers = require('./data/reporting.transfers.json');
 
-describe('GraphQL Integration Tests', () => {
+describe('TransferSummary Integration Tests', () => {
     const MONGO_URI = process.env.MONGO_URI || 'mongodb://localhost:27017';
     let collection: Collection;
 
@@ -31,6 +32,21 @@ describe('GraphQL Integration Tests', () => {
             console.error('Database cleanup failed:', error);
             throw error;
         }
+    });
+
+    describe('insert data into live MongoDB', () => {
+        it('should insert transfer summary data', async () => {
+            // Insert test data
+            const result = await collection.insertMany(testTransfers);
+
+            const insertedIds = result.insertedIds;
+
+            // get inserted ids count
+            const insertedIdsCount = Object.keys(insertedIds).length;
+
+            // Assert
+            expect(insertedIdsCount).toEqual(8);
+        });
     });
 
     describe('Send sample graphql queries', () => {
